Add utility functions for token estimation, import extraction, and file handling
